Update dependencies and modernize project structure
Chatbot: Update dependencies and modernize project structure
