+++ .gitignore
@@ -1,4 +1,3 @@
-/.qodo
 # Logs
 logs
 *.log
