--- App.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function App() {
-  return (
-    <div className="min-h-screen bg-gray-100 flex items-center justify-center">
-      <p>Start prompting (or editing) to see magic happen :)</p>
-    </div>
-  );
-}
-
-export default App;
